organization: move the pkg to a dedicated one
